Added support for getting node by unique handle
Nodes are no longer queried by their names which could be problematic in scenes with duplicate names. A unique scene handle (id) is now used instead.
For 3ds max is the nodes handle, for maya it is the nodes full name.

This has involved the implementation of a host interface class IHost.
Only implemented IHost for max so far.
This has required a large refactor of the dcc sub-package and the setup logic for skin_plus_plus itself.