Maya fixes
Fixed issue with maya extract skin data function.
When the boneIDs array was instantiated it was instantiated with a size of 0 and then resized later. The resize call did not take into account the requirement that boneID array needs to be default values of -1.
We now instantiate the array at the maximum influence count and then resize to the maximum USED influence count which in most cases will be much lower than the max influence count. This preserves the -1 defaults and reduces the array size to only what is required.
Also fixed the mel script call to add missing bones which was non-functional previously.